Remove manual pagebreaks in master file
Manual page breaks should be avoided. If it is absolutely necessary to
break the page there, one should find a semantic approach like
redefining `\section` to end with a linebreak or similiar.

But as a rule of thumb, one should just trust LaTeX to make the correct
decisions and not try to adjust things manually.